core_hub/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-environ
psycopg2-binary
razorpay
django-cors-headers==4.3.1
drf-spectacular
channels
daphne
channels-redis
whitenoise
